Move code to input_driver.c
