+++ input/input_driver.h
@@ -120,10 +120,48 @@ typedef struct
    bool autoconfigured;
 } input_device_info_t;
 
+struct remote_message
+{
+   int port;
+   int device;
+   int index;
+   int id;
+   uint16_t state;
+};
+
+struct input_remote
+{
+#if defined(HAVE_NETWORKING) && defined(HAVE_NETWORKGAMEPAD)
+   int net_fd[MAX_USERS];
+#endif
+   bool state[RARCH_BIND_LIST_END];
+};
+
+
 typedef struct
 {
    char display_name[256];
 } input_mouse_info_t;
+
+typedef struct input_remote input_remote_t;
+
+typedef struct input_remote_state
+{
+   /* This is a bitmask of (1 << key_bind_id). */
+   uint64_t buttons[MAX_USERS];
+   /* Left X, Left Y, Right X, Right Y */
+   int16_t analog[4][MAX_USERS];
+} input_remote_state_t;
+
+typedef struct input_list_element_t
+{
+   int16_t *state;
+   unsigned port;
+   unsigned device;
+   unsigned index;
+   unsigned int state_size;
+} input_list_element;
+
 
 /**
  * Organizes the functions and data structures of each driver that are accessed
@@ -671,6 +709,18 @@ int16_t input_joypad_axis(
       float input_analog_sensitivity,
       const input_device_driver_t *drv,
       unsigned port, uint32_t joyaxis, float normal_mag);
+
+#if defined(HAVE_NETWORKING) && defined(HAVE_NETWORKGAMEPAD)
+void input_remote_parse_packet(
+      input_remote_state_t *input_state,
+      struct remote_message *msg, unsigned user);
+
+input_remote_t *input_driver_init_remote(
+      settings_t *settings,
+      unsigned num_active_users);
+
+void input_remote_free(input_remote_t *handle, unsigned max_users);
+#endif
 
 /*****************************************************************************/
